requirements.txt
Flask>=1.0,<=1.1.2
azure-core==1.15.0
Flask-WTF==0.15.1
azure-ai-textanalytics==5.1.0b7
bs4==0.0.1
requests==2.25.1